{{ cookiecutter.github_repository }}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.workflow_name }}"
version = "{{ cookiecutter.workflow_version }}"
description = "{{ cookiecutter.workflow_description }}"
documentation = "{{ cookiecutter.github_url }}/blob/main/README.md"
homepage = "{{ cookiecutter.github_url }}"
license = "MIT"

authors = [
    "{{ cookiecutter.author_name }} <{{ cookiecutter.author_email }}>"
]

[tool.poetry.dependencies]
python = ">=3.8,<=3.12"
alfred-pyflow = "^0.4.4"

[tool.poetry.dev-dependencies]
black = "^23.3.0"
ddt = "^1.6.0"
coverage = "^7.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
